Add content logic

+++ backend/NetflixClone.Application/UseCases/SubscriptionPlans/Commands/DeletePlan/DeleteSubscriptionPlanHandler.cs
@@ -14,6 +14,13 @@ public class DeleteSubscriptionPlanHandler : IRequestHandler<DeleteSubscriptionP
 
     public async Task Handle(DeleteSubscriptionPlanCommand request, CancellationToken ct)
     {
+        var plan = await _repository.GetByIdAsync(request.Id, ct);
+
+        if (plan == null)
+        {
+            throw new Exception("Підписки не знайдено.");
+        }
+
         await _repository.DeleteAsync(request.Id, ct);
     }
 }
